ch4/requirements.txt
absl-py==0.4.1
astor==0.7.1
blocksparse==1.0.0
click==6.7
cloudpickle==0.5.5
cycler==0.10.0
dask==0.19.1
decorator==4.3.0
dlib==19.15.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
gast==0.2.0
grpcio==1.14.2
ImageHash==4.0
imageio==2.4.1
imutils==0.5.1
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
networkx==2.1
numpy==1.14.5
opencv-python==3.4.2.17
pandas==0.23.4
Pillow==5.2.0
pkg-resources==0.0.0
protobuf==3.6.1
pyparsing==2.2.0
python-dateutil==2.7.3
python-slugify==1.2.6
pytz==2018.5
PyWavelets==1.0.0
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
toolz==0.9.0
tqdm==4.25.0
Unidecode==1.0.22
Werkzeug==0.14.1
